+++ app/electron-builder.json
@@ -1,0 +1,39 @@
+{
+  "appId": "org.enso.studio",
+  "productName": "Enso Studio",
+  "copyright": "Copyright © 2020 ${author}.",
+  "mac": {
+    "icon": "dist/icon/icon.icns",
+    "category": "public.app-category.developer-tools",
+    "darkModeSupport": true,
+    "type": "distribution"
+  },
+  "win": {
+    "icon": "dist/icon/icon.ico"
+  },
+  "linux": {
+    "icon": "dist/icon/png",
+    "category": "Development"
+  },
+  "files": [
+    "dist/assets/**/*",
+    "dist/main/**/*",
+    "dist/renderer/**/*",
+    "package.json"
+  ],
+  "fileAssociations": [
+    {
+      "ext": "enso",
+      "name": "Enso Source File",
+      "role": "Editor"
+    },
+    {
+      "ext": "enso-studio",
+      "name": "Enso Studio Project",
+      "role": "Editor"
+    }
+  ],
+  "directories": {
+    "output": "dist/native"
+  }
+}
