+++ app/electron-webpack.json
@@ -1,0 +1,8 @@
+{
+  "main": {
+    "sourceDirectory": "src/bin/native"
+  },
+  "renderer": {
+    "sourceDirectory": null
+  }
+}
