+++ app/src/lib/animation.js
